fix(skill-sync): break custom skills import cycle
